Functional iOS Emulator and Native Installs
Installed pods to allow for the launching and installing on the iOS emulator and natively installing. I also changed the target version of iOS from 11 to 13.4, but that can be changed easily if needed.

Note:
To launch via Xcode (which now works because the emulator works), open the ShelfLifeNative.xcworkspace file, NOT ShelfLifeNative.xcodeproj, and you can hit the play button there.