Added support for going to next screen
This currently passes in all the data for the selected recipe, so there is no need to sort through the entire JSON file. Additionally, the data is passed pretty manually, so if data is added or removed from every objects (perhaps a steps string or list), it'd have to be added to (currently) line 41.